+++ pressed/src/main/scala/de/welt/contentapi/pressed/models/ApiConfiguration.scala
@@ -86,18 +86,20 @@ case class ApiCommercialTaboolaConfiguration(showNews: Option[Boolean] = None,
 /**
   * Sponsoring of section and content pages. The impl is part of the client.
   *
-  * @param name   name of the branding. Need for mapping.
-  * @param logo   name of the sponsoring logo. Need for mapping. Not a real image path.
-  * @param slogan optional slogan of the sponsoring.
-  * @param hidden name of the branding. Need for mapping.
-  * @param link   Optional link for the logo.
+  * @param name         name of the branding. Need for mapping.
+  * @param logo         name of the sponsoring logo. Need for mapping. Not a real image path.
+  * @param slogan       optional slogan of the sponsoring.
+  * @param hidden       name of the branding. Need for mapping.
+  * @param link         Optional link for the logo.
+  * @param brandstation Optional brandstation type. Need for mapping.
   */
 case class ApiSponsoringConfiguration(@deprecated("Renaming. Use logo instead", since = "01/2017")
                                       name: Option[String] = None,
                                       logo: Option[String] = None,
                                       slogan: Option[String] = None,
                                       hidden: Option[Boolean] = None,
-                                      link: Option[ApiReference] = None) {
+                                      link: Option[ApiReference] = None,
+                                      brandstation: Option[String] = None) {
   lazy val isHidden: Boolean = hidden.getOrElse(false)
 }